migrate building elf file to ELF class from Assembler class
