+++ lib/vaporware/compiler/assembler.rb
@@ -10,9 +10,8 @@ class Vaporware::Compiler::Assembler
 
   def initialize(input:, output: File.basename(input, ".*") + ".o", assembler: "as", type: :relocatable, debug: false)
     @input, @output = input, output
-    @elf_header = ELF::Header.new(type:)
+    @elf = ELF.new(type:, input:, output:, debug:)
     @assembler = assembler
-    @sections = ELF::Sections.new
     @debug = debug
   end
 
@@ -20,71 +19,10 @@ class Vaporware::Compiler::Assembler
     if ["gcc", "as"].include?(assembler)
       IO.popen([assembler, *assembler_options, "-o", output, input].join(" ")).close
     else
-      to_elf(input:, output:)
+      to_elf(input:, output:, debug:)
     end
     output
   end
   def obj_file = @output
-
-  def to_elf(input: @input, output: @output, debug: false)
-    program_size = 0
-    read!(input:)
-    init_assemble!
-
-    offset = 0x40
-    section_headers = []
-    names = []
-    bodies = {
-      null: nil,
-      text: nil,
-      data: nil,
-      bss: nil,
-      note: nil,
-      symtab: nil,
-      strtab: nil,
-      shstrtab: nil,
-    }
-    name_idx = 0
-    padding = nil
-    @sections.each do |section|
-      name = section.name
-      names << name
-      section.body.set!(name: names.join) if name == "\0.shstrtab"
-      bin = section.body.build
-      size = bin.bytesize
-      bin << "\0" until (bin.bytesize % 8) == 0 if ["\0.text", "\0.shstrtab"].include?(name)
-      bin << "\0" until ((bin.bytesize + offset) % 8) == 0 if ["\0.shstrtab"].include?(name)
-      bodies[section.section_name.to_sym] = bin
-      header = section.header
-      if offset > 0x40 && size > 0 && padding&.>(0)
-        offset += padding
-        padding = nil
-      end
-      padding = bin.size - size if size > 0
-      header.set!(name: name_idx, offset:, size:) unless name == ""
-      offset += size
-      section_headers << header.build
-      name_idx += name == "" ? 1 : name.size
-    end
-    @elf_header.set!(shoffset: offset + padding)
-    w = File.open(output, "wb")
-    w.write([@elf_header.build, *bodies.values, *section_headers].join)
-    w.close
-    [@elf_header.build, *bodies.values, *section_headers]
-  end
-
-  private
-  def init_assemble! = (note!; symtab!)
-  def read!(input: @input, text: @sections.text.body)
-    read = { main: false }
-    File.open(input, "r") do |r|
-      r.each_line do |line|
-        read[:main] = line.match(/main:/) unless read[:main]
-        next unless read[:main] && !/main:/.match(line)
-        text.assemble!(line)
-      end
-    end
-  end
-  def note! = @sections.note.body.null!
-  def symtab! = @sections.symtab.body.set!(entsize: 0x18, name: 1, info: 0x10, other: 0, shndx: 1)
+  def to_elf(input: @input, output: @output, debug: false) = @elf.build(input:, output:, debug:)
 end
