+++ sig/vaporware/compiler/assembler.rbs
@@ -1,25 +1,13 @@
-use Vaporware::Compiler::Assembler::ELF::Header
-use Vaporware::Compiler::Assembler::ELF::Sections
-use Vaporware::Compiler::Assembler::ELF::Section::Text
-
+use Vaporware::Compiler::Assembler::ELF
 class Vaporware::Compiler::Assembler
   @input: String
   @output: String
   @assembler: String
-  @elf_header: Header
-  @sections: Sections
+  @elf: ELF
   @debug: bool
 
   def initialize: (input: String, ?output: String, ?type: Symbol, ?debug: bool) -> void
   def assemble: (?assembler: String, ?assembler_options: Array[String] | [], ?input: String, ?output: String, ?debug: bool) -> String
   def obj_file: () -> String
   def to_elf: (?input: String, ?output: String, ?debug: bool) -> void
-  
-  private def read!: (?input: String, ?text: Text) -> void
-  private def init_assemble!: () -> void
-  private def note!: () -> void
-  private def text!: () -> void
-  private def symtab!: () -> void
-  private def strtab!: () -> void
-  private def shstrtab!: () -> void
 end
